Cleaning rosmaster code
